feature: Pp.paragraph
We add two new Pp helpers for building paragraphs of text. This
surrounds Pp.text in a Pp.hovbox which allows it to be used freely
without being distored by a surrounding Pp.vbox which is a common
occurance.

We demonstrate the use with a test contrasting it with the other ways to
write paragraphs.

Signed-off-by: Ali Caglayan <alizter@gmail.com>